refactor: Main클래스 생성 및 Race클래스의 사용자 입출력 메소드들을 이동
